+++ src/cli/assets/platform.suite.template.js
@@ -1,10 +1,11 @@
 'use strict';
 
-const tester = require('core/tester');
-const schema = require('./assets/%name.schema.json');
+const suite = require('core/suite');
+const payload = require('./assets/%name');
+const schema = require('./assets/%name.schema');
 
-tester.for(null, '%name', (api) => {
-  // checkout functions available under tester.test which provide a lot of pre-canned tests
+suite.forPlatform('%name', schema, payload, (test) => {
+  // checkout functions available under test.should which provide a lot of pre-canned tests
   //   more information here: https://github.com/cloud-elements/churros/blob/master/CONTRIBUTING.md#adding-tests-to-an-existing-suite
 
   it('%user should insert some tests here :)', () => true);