+++ src/test/elements/amazons3/files.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const tester = require('core/tester');
-
-describe('files', () => {
-  tester.for('documents', 'files');
-});