+++ src/test/elements/closeio/accounts.js
@@ -1,17 +1,9 @@
 'use strict';
 
-const tools = require('core/tools');
-const tester = require('core/tester');
+const suite = require('core/suite');
 const schema = require('./assets/account.schema');
+const account = require('./assets/account');
 
-const gen = (opts) => {
-  opts = opts ? opts : {};
-  const random = tools.random();
-  return new Object({
-    name: (opts.name || 'mr. churros ' + random)
-  });
-};
-
-tester.for('crm', 'accounts', (api) => {
-  tester.test.cruds(api, gen(), schema);
+suite.forElement('crm', 'accounts', account, schema, (test) => {
+  test.should.supportCruds();
 });