+++ src/test/elements/shopify/orders.js
@@ -1,19 +1,16 @@
 'use strict';
 
-const tester = require('core/tester');
+const suite = require('core/suite');
 const tools = require('core/tools');
 const schema = require('./assets/orders.schema');
 
-
 const order = (custom) => new Object({
-  line_items: custom.line_items || [
-    {
-      title: tools.random(),
-      price: tools.randomInt()
-    }
-  ]
+  line_items: custom.line_items || [{
+    title: tools.random(),
+    price: tools.randomInt()
+  }]
 });
 
-tester.for('ecommerce', 'orders', (api) => {
-  tester.test.cruds(api, order({}), schema);
+suite.forElement('ecommerce', 'orders', order({}), schema, (test) => {
+  test.should.supportCruds();
 });