+++ src/test/elements/shopify/query.js
@@ -1,22 +1,9 @@
 'use strict';
 
-const tester = require('core/tester');
+const suite = require('core/suite');
 const schema = require('./assets/customers.schema');
 
-tester.for('ecommerce', 'query', (api) => {
-
-  it('should allow a query for all fields', () => {
-    var apiWithQuery = api;
-    apiWithQuery += '?q=';
-    apiWithQuery += encodeURIComponent('select * from customers');
-    return tester.find(apiWithQuery, schema);
-  });
-
-  it('should allow a query for specific fields', () => {
-    var apiWithQuery = api;
-    apiWithQuery += '?q=';
-    apiWithQuery += encodeURIComponent('select id, email from customers');
-    return tester.find(apiWithQuery, schema);
-  });
-
+suite.forElement('ecommerce', 'query', null, schema, (test) => {
+  test.withOptions({ qs: { q: 'select * from customers' } }).should.return200OnGet();
+  test.withOptions({ qs: { q: 'select id, email from customers' } }).should.return200OnGet();
 });