+++ src/test/elements/shopify/shops.js
@@ -1,11 +1,8 @@
 'use strict';
 
-const tester = require('core/tester');
+const suite = require('core/suite');
 const schema = require('./assets/shops.schema');
 
-
-tester.for('ecommerce', 'shops', (api) => {
-  it('should allow GET for /hubs/ecommerce/shops', () => {
-    return tester.find(api, schema);
-  });
+suite.forElement('ecommerce', 'shops', null, schema, (test) => {
+  test.should.return200OnGet();
 });