+++ src/test/elements/taxify/taxes.js
@@ -1,12 +1,12 @@
 'use strict';
 
-const tester = require('core/tester');
+const suite = require('core/suite');
 const schema = require('./assets/taxes.schema');
 const payload = require('./assets/taxes.create');
 const badPayload = require('./assets/taxes.bad.create');
 
-tester.for('finance', 'taxes', (api) => {
-  tester.test.create(api, payload, schema);
-  tester.test.badPost400(api, badPayload);
-  tester.test.badPost400(api);
+suite.forElement('finance', 'taxes', payload, schema, (test) => {
+  test.should.return200OnPost();
+  test.withJson(badPayload).should.return400OnPost();
+  test.withJson({}).should.return400OnPost();
 });