+++ src/test/platform/formulas/formulas.js
@@ -3,7 +3,8 @@
 const provisioner = require('core/provisioner');
 const util = require('util');
 const tools = require('core/tools');
-const tester = require('core/tester');
+const suite = require('core/suite');
+const cloud = require('core/cloud');
 const chakram = require('chakram');
 const expect = chakram.expect;
 const schema = require('./assets/formula.schema');
@@ -26,31 +27,32 @@ const genInstance = (opts) => new Object({
   name: (opts.name || 'churros-formula-instance-name')
 });
 
-tester.for(null, 'formulas', (api) => {
-  tester.test.crud(api, genFormula({}), schema, chakram.put);
+suite.forPlatform('formulas', genFormula({}), schema, (test) => {
+
+  test.should.supportCrud(chakram.put);
 
   it('should allow adding and removing "scheduled" trigger to a formula', () => {
     const f = genFormula({});
     const t = genTrigger({});
 
     let formulaId;
-    return tester.post(api, f, schema)
+    return cloud.post(test.api, f, schema)
       .then(r => {
         formulaId = r.body.id;
-        const url = util.format('%s/%s/triggers', api, formulaId);
-        return tester.post(url, t, triggerSchema);
+        const url = util.format('%s/%s/triggers', test.api, formulaId);
+        return cloud.post(url, t, triggerSchema);
       })
       .then(r => {
-        const url = util.format('%s/%s/triggers/%s', api, formulaId, r.body.id);
-        return tester.get(url, triggerSchema);
+        const url = util.format('%s/%s/triggers/%s', test.api, formulaId, r.body.id);
+        return cloud.get(url, triggerSchema);
       })
       .then(r => {
         const url = util.format('/formulas/%s/triggers/%s', formulaId, r.body.id);
-        return tester.delete(url);
+        return cloud.delete(url);
       })
       .then(r => {
         const url = util.format('/formulas/%s', formulaId);
-        return tester.delete(url);
+        return cloud.delete(url);
       });
   });
 
@@ -58,7 +60,7 @@ tester.for(null, 'formulas', (api) => {
     const f = genFormula({});
 
     let formulaId;
-    return tester.post(api, f, schema)
+    return cloud.post(test.api, f, schema)
       .then(r => {
         formulaId = r.body.id;
 
@@ -67,32 +69,32 @@ tester.for(null, 'formulas', (api) => {
             cron: '0 0/14 * 1/1 * ? *'
           }
         });
-        const tApi = util.format('%s/%s/triggers', api, formulaId);
+        const tApi = util.format('%s/%s/triggers', test.api, formulaId);
 
-        return tester.post(tApi, t, (r) => expect(r).to.have.statusCode(400));
+        return cloud.post(tApi, t, (r) => expect(r).to.have.statusCode(400));
       })
-      .then(r => tester.delete(api + '/' + formulaId));
+      .then(r => cloud.delete(test.api + '/' + formulaId));
   });
 
   it('should not allow creating an instance of a formula with an invalid on success step', () => {
     const f = genFormula({});
 
     let formulaId;
-    return tester.post(api, f, schema)
+    return cloud.post(test.api, f, schema)
       .then(r => {
         formulaId = r.body.id;
         const t = genTrigger({
           onSuccess: ['fake-step-name']
         });
-        const tApi = util.format('%s/%s/triggers', api, formulaId);
-        return tester.post(tApi, t, triggerSchema);
+        const tApi = util.format('%s/%s/triggers', test.api, formulaId);
+        return cloud.post(tApi, t, triggerSchema);
       })
       .then(r => {
         const fi = genInstance({});
         const iApi = util.format('/formulas/%s/instances', formulaId);
-        return tester.post(iApi, fi, (r) => expect(r).to.have.statusCode(400));
+        return cloud.post(iApi, fi, (r) => expect(r).to.have.statusCode(400));
       })
-      .then(r => tester.delete(util.format('/formulas/%s', formulaId)));
+      .then(r => cloud.delete(util.format('/formulas/%s', formulaId)));
   });
 
   it('should allow creating a big azz formula and then an instance', () => {
@@ -107,13 +109,13 @@ tester.for(null, 'formulas', (api) => {
         f.name = tools.random();
 
         let formulaId;
-        return tester.post(api, f, schema)
+        return cloud.post(test.api, f, schema)
           .then(r => {
             formulaId = r.body.id;
-            return tester.post(util.format('/formulas/%s/instances', formulaId), fi, instanceSchema);
+            return cloud.post(util.format('/formulas/%s/instances', formulaId), fi, instanceSchema);
           })
-          .then(r => tester.delete(util.format('/formulas/%s/instances/%s', formulaId, r.body.id)))
-          .then(r => tester.delete(util.format('/formulas/%s', formulaId)))
+          .then(r => cloud.delete(util.format('/formulas/%s/instances/%s', formulaId, r.body.id)))
+          .then(r => cloud.delete(util.format('/formulas/%s', formulaId)))
           .then(r => provisioner.delete(id));
       });
   });