+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v0.2.0 <sub><sup>(TBD)</sup></sub>
+## v0.2.0 <sub><sup>(2016-02-12)</sup></sub>
 
 #### Highlights
 * Added a large amount of new elements to our provisioning library.  Most of the ones left were OAuth1 and OAuth2 elements, so lots of selenium work :disappointed:
@@ -9,6 +9,7 @@
 * Support for `--verbose` mode when running `churros test`
 * Prompt for authentication type and properties when calling `churros add element`
 * Added optional `--browser <name>` support so, if you're debugging your OAuth element, you can now actually *see* the browser running your selenium commands.  (i.e. `churros test elements/foo --browser firefox`)
+* Drastically changed the way you interact with tests inside what was formally known as `tester.for`, and is now: `suite.forElement` or `suite.forPlatform`
 
 ## v0.1.0 <sub><sup>(2016-1-29)</sup></sub>
 
@@ -18,7 +19,7 @@
 * Continuing to add more elements to the elements catalog
 * Added support for `should.have.schemaAnd200` `chakram` assertion as I found that I was always doing both of those at the same time and also because the error message for each of those wasn't good enough so wanted to improve it
 * Added support for `should.have.statusCode` which is the *exact* same as `should.have.status` but gives more logging when the assertion fails to help debug the issue more easily.
-* Many common tests available under `core/tester` and can be found under the `tester.test.` namespace.  This library will continue to grow as we add more and more tests and find out the commonalities throughout more elements.
+* Many common tests available under `core/suite` and can be found under the `suite.it.` namespace.  This library will continue to grow as we add more and more tests and find out the commonalities throughout more elements.
 * Event tests that create an instance of an element with notifications enabled, register a local URL that is exposed (can use ngrok, localtunnel, ssh tunneling) and then listen on that URL for incoming events from the platform.  Also supports specifying how many events to send and how long to wait to receive the webhook notification.
 * `churros add` sub-command to help setup a new platform or element suite